+++ NEWS.md
@@ -1,3 +1,7 @@
+# oeli development
+
+* New function `function_arguments()` that returns the names of function arguments.
+
 # oeli 0.1.0
 
 * Initial CRAN submission.